+++ Prendas_Server/src/entities/ItemRecetaEntity.java
@@ -13,7 +13,7 @@ public class ItemRecetaEntity implements Serializable{
 	@Id
 	@GeneratedValue
 	private int id;
-	private float cantidad;
+	private double cantidad;
 	private double desperdicio;
 	@OneToOne(fetch=FetchType.EAGER)
 	private InsumoEntity insumo;
@@ -26,10 +26,10 @@ public class ItemRecetaEntity implements Serializable{
 	public void setId(int id) {
 		this.id = id;
 	}
-	public float getCantidad() {
+	public double getCantidad() {
 		return cantidad;
 	}
-	public void setCantidad(float cantidad) {
+	public void setCantidad(double cantidad) {
 		this.cantidad = cantidad;
 	}
 	public double getDesperdicio() {
